XXX Convert Comb and PalColour to work on fields, not frames.
For PalColour, this was straightforward, since it already worked on
fields and just interlaced the result at the end.

For Comb, it's reasonably simple, but the 3D filter now needs a
two-field delay rather than a one-frame delay. I've converted a couple
of places that used 1-based indexing to 0-based indexing.

The helper function in Decoder now interlaces the results to produce the
final frame.

OpticalFlow similarly needs a two-field delay, and the horizontal
sensitivity adjustment isn't needed any more.

The code is overall a bit simpler, and reveals a bit more about what
information the decoders actually need about each field. It produces the
same results as the previous version.

XXX update ld-analyse (or provide a better interface for it)