Move most of Comb's methods into Comb::FrameBuffer.
Nearly all of the decoding steps are operating on a single frame, so
they can be a method of FrameBuffer, rather than a method of Comb that
takes a FrameBuffer as a parameter. This simplifies the code quite a bit
because it doesn't need to say currentFrameBuffer->... all over the
place.

The code to load two fields into the framebuffer is also made into a
method, and the Get* helpers are renamed to get*.