Fix a typo in the split2D heuristic.
This survives from the original version of this code; it looks like the
intent was to use the current and previous lines, not the current line
twice.

Because they're usually very similar, the difference to the result is
tiny -- the cross-colour in the S&W zoneplate is slightly better in a
couple of pixels.